Parte de desconto 3x2 (falta gardar na base de datos os articulos dos packs)

+++ ProyectoWI/src/java/tienda/Helpers/UsuarioHelper.java
@@ -169,7 +169,7 @@ public class UsuarioHelper {
                 request.getParameter("email"),
                 activo,
                 Integer.parseInt(request.getParameter("idTipoUsuario")),
-                Float.parseFloat(request.getParameter("totalCompra")));
+                Double.parseDouble(request.getParameter("totalCompra")));
         String idDireccionS = request.getParameter("idDireccion");
         Integer idDireccion = null;
         if (idDireccionS != null) {
